revert(views): refactor module;
- Reverted module rename.
